Conflicts:
	Clap/src/main/java/shopping/controller/DoCheckOutAction.java
	Clap/src/main/webapp/sql/SQLQuery2.sql

+++ Clap/src/test/java/shopping/model/TestOrderFormService.java
@@ -32,7 +32,7 @@ public class TestOrderFormService extends TestCase {
 	private Integer id = 36;
 	private String mb_email = "lee@gmail.com";
 	private Timestamp time;
-	private Integer total = 9999;
+	private Double	total = 9999.0;
 	private Byte status;
 
 	@Before
